requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
future==1.0.0
iso8601==2.1.0
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
scipy==1.13.0
six==1.16.0
statsmodels==0.14.2
tzdata==2024.1
